Move up testing project
