+++ Lombiq.ContentEditors.Tests.UI/License.md
