+++ Lombiq.ContentEditors.Tests.UI/NuGetIcon.png
