--- Lombiq.ContentEditors/Tests/Lombiq.ContentEditors.Tests.UI/Lombiq.ContentEditors.Tests.UI.csproj
@@ -1,32 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <Title>Lombiq Content Editors for Orchard Core - UI Test Extensions</Title>
-    <Authors>Lombiq Technologies</Authors>
-    <Copyright>Copyright © 2022, Lombiq Technologies Ltd.</Copyright>
-    <Description>Lombiq Content Editors for Orchard Core - UI Test Extensions: Extensions to aid in UI testing Lombiq Content Editors for Orchard Core.</Description>
-    <PackageIcon>NuGetIcon.png</PackageIcon>
-    <PackageTags>OrchardCore;Lombiq;AspNetCore</PackageTags>
-    <RepositoryUrl>https://github.com/Lombiq/Orchard-Content-Editors</RepositoryUrl>
-    <PackageProjectUrl>https://github.com/Lombiq/Orchard-Content-Editors</PackageProjectUrl>
-    <PackageLicenseFile>License.md</PackageLicenseFile>
-  </PropertyGroup>
-
-  <ItemGroup Condition="'$(NuGetBuild)' != 'true'">
-    <ProjectReference Include="..\..\..\..\..\..\test\Lombiq.UITestingToolbox\Lombiq.Tests.UI\Lombiq.Tests.UI.csproj" />
-  </ItemGroup>
-
-  <ItemGroup Condition="'$(NuGetBuild)' == 'true'">
-    <PackageReference Include="Lombiq.Tests.UI" Version="6.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="License.md" Pack="true" PackagePath="" />
-    <None Include="NuGetIcon.png" Pack="true" PackagePath="" />
-  </ItemGroup>
-
-</Project>
